ep ready for staging :dragon:

+++ backend/src/routes.php
@@ -10,27 +10,6 @@ $app->get('/home', function ($request, $response, $args) {
     // Render index view
     return $this->renderer->render($response, 'index.html', $args);
 });
-$app->get('/dbcse3330',function($request,$response,$args){
-
-	$todo=$this->dbcse3330;
-	$str='';
-	foreach($todo->query('select * from todo where id=0') as $row){
-		    echo "<tr>";
-        foreach($row as $key => $value){
-            echo "<td>" . $value . "</td>"."\n";
-        }
-         echo "</tr>";
-	}
-	//return $response->write($str);
-});
-$app->get('/m',function($request,$response,$args){
-echo $args['x'];
-echo $args['y'];
-// In case any of our lines are larger than 70 characters, we should use wordwrap()
-// Send
-
-	//return $response->write($str);
-});
 $app->get('/admin/getTutors',function($request,$response,$args){
   //$query = mysql_query("SELECT * FROM Tutor");
       $db = $this->alec;
@@ -68,8 +47,16 @@ $app->get('/admin/getTutors/{id}',function($request,$response,$args){
 });
 $app->delete('/admin/getTutors/{id}',function($request,$response,$args){
   //$query = mysql_query("SELECT * FROM Tutor");
+
         $id=$args['id'];
+        if(!isset($id))
+        {
+          $new_response=$response->withStatus(400);
+          echo "please provide a user id";
+          return $new_response;
+        }
         $db = $this->alec;
-         $result=$db->query('delete from Tutor where id='.$id);
-echo 'success';
+        $result=$db->query('delete from Tutor where id='.$id);
+        echo 'success';
+   return $response;
 });
